Create teacher refactor
