+++ app.rb
@@ -55,19 +55,6 @@ class App
     option_checker(answer)
   end
 
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Specialization: '
-    specialization = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-
-    teacher = Teacher.new(age, specialization, name)
-    @people.push(teacher)
-    puts 'Teacher is created successfully'
-  end
-
   def create_book
     print 'Title: '
     title = gets.chomp
